refactor: (In progress) 액터-WorldPartition 상호작용 배제 중
- UStaticMeshComponent 메소드 이름 변경: MarkBVDirty() -> MarkWorldPartitionDirty()

- AActor 외부에서 AActor::MarkPartitionDirty() 제거. (RootComponent의 OnTransformUpdated로 대체)

- OnTransformUpdated에 템플릿 메소드 패턴 적용. OnTransformUpdated는 공통로직(Propagation 등)만 수행하고 클래스별 세부 구현은 OnTransformUpdatedChildImpl() 가상함수 이용

+++ TL2/DecalComponent.cpp
@@ -28,7 +28,7 @@ void UDecalComponent::DuplicateSubObjects()
 void UDecalComponent::TickComponent(float DeltaTime)
 {
 	// Opacity 업데이트
-	DecalOpacity += FadeDirection * FadeSpeed * DeltaTime;
+	DecalOpacity += static_cast<float>(FadeDirection) * FadeSpeed * DeltaTime;
 
 	// 범위 체크 및 방향 반전
 	if (DecalOpacity <= 0.0f)
